+++ pr1/ITask.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using pr1;
+
+namespace pr1
+{
+    interface ITask
+    {
+        void Execute();
+    }
+}
